requirements.txt
streamlit==1.31.1
requests==2.31.0
Pillow==10.2.0
openai==1.12.0
